Sel4: Performance comparisons
Bring int the actual device tree source (.dts) that gets compile
as a .dtb inside the camkes-arm-vm project[1] from sel4. The device
tree sources in that project are different from the ones in this
linux repository.

This is an attempt to get a matching kernel+device-tree to compare the
performance between the virtualized linux client on the Sel4 hypervisor
vs. the same linux kernel+device-tree running natively on a Tegra TK1.

[1]: https://github.com/SEL4PROJ/camkes-arm-vm

Signed-off-by: Jorge Sanjuan <jorge.sanjuan@codethink.co.uk>