DTS: Command line: Limit system memory to 1GB
When running Linux as a client on the sel4 hypervisor, the system
RAM memory is limited to only 1GB. We limit this from comman line
as it looks U-boot is setting up the memory node to tell the
real memory on the system. Hence, changes on this node in Device
Tree are not limiting the memory as expected.

Signed-off-by: Jorge Sanjuan <jorge.sanjuan@codethink.co.uk>